// src/components/showAlert/types.ts
export type BgStyle = 
  | 'bg-red-500'
  | 'bg-white'
  | 'bg-yellow-500'
  | 'bg-green-500'
  | 'bg-transparent';

export type TextStyle = 
  | 'text-red-500'
  | 'text-white'
  | 'text-yellow-500'
  | 'text-green-500'
  | 'text-black'; 

export type BorderStyle =
  | 'border-red-500'
  | 'border-white'
  | 'border-yellow-500'
  | 'border-green-500'
  | 'border-black'; 